requirements.txt
cssmin==0.2.0
django-htmlmin==0.11.0
Flask==0.12.4
Flask-Assets==0.12
Flask-Compress==1.4.0
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
nose==1.3.7
psycopg2-binary==2.8.2
pyquery==1.4.0
pyScss==1.3.5
python-decouple==3.1
PyYAML==5.1
selenium==3.141.0
SQLAlchemy==1.3.3
Unipath==1.1